net/ice: use correct format specifiers for unsigned ints
Firmware was giving a number for the MSIX vectors that was
way too big and obviously not right. Because of the wrong
format specifier, this big number ended up looking like a
tiny negative number in the logs. This was fixed by using
the right format specifier everywhere it's needed.

Signed-off-by: Yogesh Bhosale <yogesh.bhosale@intel.com>
Signed-off-by: Soumyadeep Hore <soumyadeep.hore@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>