bus/fslmc/dpaa2: replace system("echo ...") with file i/o
Using system() is a bad idea in driver code because it introduces
a number of potential security issues. The codeql analysis tool
flags this a potential security issue.

Instead just use normal stdio to do the same thing.

Compile test only, do not have this hardware and therefore can
not test this.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Reviewed-by: Sachin Saxena <sachin.saxena@oss.nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>