net/ixgbe/base: remove minsrevs code from DPDK
Commit add44414762c ("net/ixgbe/base: add E610 NVM-related operations") has
added code to read minimum security revision from NVM. This code was not
meant to be included in DPDK, and was only meant for other drivers derived
from shared base code, but was present due to the way shared driver code
snapshot was generated. Remove this code from DPDK driver.

Fixes: add44414762c ("net/ixgbe/base: add E610 NVM-related operations")

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>