net/nfp: reserve BAR for expansion ROM
For some platform, the warm restart of host doesn't trigger
the reset of NIC, which causes the initialize process of NIC
not executed with the right expansion ROM mapping. Consequently,
the PXE boot won't work in this case.

Now reserve BAR 2.0 which used by expansion ROM so that the
mapping is fixed.

Signed-off-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Peng Zhang <peng.zhang@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>