devtools/test-meson-builds: use cross file for 32bit build
When testing the 32-bit x86 build on debian or ubuntu linux systems, use
the cross-file rather than using args and pkgconfig environment
variable. The advantage of using the cross-file is that the paths are
saved across runs. While the '-m32' args settings are preserved in the
current setup, the PKG_CONFIG_LIBDIR value from environment is not,
which can cause rebuilds of the build-32b directory to fail if meson
needs to do a reconfiguration first.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>