vfio: check iova if already mapped before do map
If we map two continuous memory area A and B, current implementation
will merge these two segments into one, as area C. But, if area A and
B are mapped again, after sort, there while be A, C, B in mem maps,
as A and B divided by C, these segs couldn't be merged. In other words,
if segments A and B that are adjacent, are mapped twice, there while be
two map entries Corresponding to A or B.

So when we partially unmap adjacent area A and B, entry C will be Residual
within mem maps. Then map an other memory area D which size is different
with A, but within area C, the area C while be mistakenly found by
find_user_mem_maps when unmapping area D. As area D and area C are of
different chunk size, this resulted in failed to unmap area D.

Fix this by check iova if already mapped before do dma map, if iova is
absolutely within mem maps, return whithout vfio map, while iova is
overlapping with entry in mem maps, return -EEXISTS.

Fixes: 56259f7fc010 ("vfio: allow partially unmapping adjacent memory")
Cc: stable@dpdk.org

Signed-off-by: Lipei Liang <lianglipei@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>