pcapng: fix handling of chained mbufs
The pcapng generates corrupted files when dealing with chained mbufs.
This issue arises because in rte_pcapng_copy the length of the EPB block
is incorrectly calculated using the data_len of the first mbuf instead
of the pkt_len, despite that rte_pcapng_write_packets correctly writing
the mbuf chain to disk.

This fix ensures that the block length is calculated based on the pkt_len,
aligning it with the actual data written to disk.

Signed-off-by: Oleksandr Nahnybida <oleksandrn@interfacemasters.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>