mempool: obey configured cache size
Seeking feedback on the concept.
I have not yet benchmarked performance.

The mempool cache size is configurable, but it could hold 1.5 times the
configured size.
This was confusing for developers, and added complexity when configuring
mempools with caches.

This patch modifies the mempool cache to obey the configured size, and
removes the cache flush threshold.

Furthermore, the mempool caches are now completely flushed/filled to/from
the backend, so backend accesses are CPU cache line aligned.

Finallly, the mempool get and put functions are optimized to only
inline the likely scenarios, and call a non-inline static helper function
in other cases.

Variuos drivers accessing the mempool directly have been updated
accordingly.

Signed-off-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>