vhost: add logging mechanism for reconnection
This patch introduces a way for backend to keep track
of the needed information to be able to reconnect without
frontend cooperation.

It will be used for VDUSE, which does not provide interface
for the backend to save and later recover local virtqueues
metadata needed to reconnect.

Vhost-user support could also be added for improved packed
ring reconnection support.

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Reviewed-by: Chenbo Xia <chenbox@nvidia.com>
Reviewed-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>