app/crypto-perf: test queue pair priority
Updated the application to test variable queue pair priority.
A mask using `--low-prio-qp-mask` can be set to lower the priority
of queues which are set in the mask.
This would result in lower performance for those queues.
By default the priority is set as highest.
This option is added just to verify the performance drop
of queues which have lower priority set.

Signed-off-by: Akhil Goyal <gakhil@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>