app/test: refactor cryptodev test cases
This commit introduces several changes to the cryptodev
test cases that should make it easier to maintain.

Changes included in this patch:
- If not needed by the specific test case, the device should be
started/stopped in the particular testsuite setup/teardown
function.
- Most of the remaining test vectors were moved from test.c file
to the specific header vector files.
- Part of GCM redundant functions were replaced by named test cases.
- Unit tests do not need to check for the symmetric cryptography feature,
if this feature were not present, the test should not even reach this stage.

Signed-off-by: Arkadiusz Kusztal <arkadiuszx.kusztal@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>